Bug 1966196 - Expand EnvironmentCoordinate hops bitwidth to 16 r=nbp
16 bits allows JSOp::EnvCallee to support more than 255 hops, but that means
GetAliasedVar, GetAliasedDebugVar, SetAliasedVar, InitAliasedLexical, and
CheckAliasedLexical also increase their length to make hops 16 bits.

Tests breakpoint-dot-generator.js and bug1368736.js needed to adjust their
bytecode offsets to mach the new operation bytelengths.

Test super-in-nested-eval.js no longer runs into the limit on environment chain
lengths (which was 255 and is now 65535), so I set the maximum depth to 300 to
show the 255 limit can now be broken, and verified that large enough depth will
throw InternalError, rather than crashing, by using a ridiculously large
number.

Differential Revision: https://phabricator.services.mozilla.com/D270389

UltraBlame original commit: 477b8f5ef6b44a96d8e286fbd48884667b9a70aa

+++ js/src/frontend/NameAnalysisTypes.h
@@ -281,7 +281,7 @@ class NameLocation {
   
   
   
-  uint8_t hops_;
+  uint16_t hops_;
 
   
   
@@ -294,7 +294,7 @@ class NameLocation {
   static_assert(LOCALNO_BITS == ENVCOORD_SLOT_BITS,
                 "Frame and environment slots must be same sized.");
 
-  NameLocation(Kind kind, BindingKind bindingKind, uint8_t hops = UINT8_MAX,
+  NameLocation(Kind kind, BindingKind bindingKind, uint16_t hops = UINT16_MAX,
                uint32_t slot = 0)
       : kind_(kind), bindingKind_(bindingKind), hops_(hops), slot_(slot) {}
 
@@ -330,13 +330,13 @@ class NameLocation {
     return NameLocation(Kind::FrameSlot, bindKind, 0, slot);
   }
 
-  static NameLocation EnvironmentCoordinate(BindingKind bindKind, uint8_t hops,
+  static NameLocation EnvironmentCoordinate(BindingKind bindKind, uint16_t hops,
                                             uint32_t slot) {
     MOZ_ASSERT(slot < ENVCOORD_SLOT_LIMIT);
     return NameLocation(Kind::EnvironmentCoordinate, bindKind, hops, slot);
   }
   static NameLocation DebugEnvironmentCoordinate(BindingKind bindKind,
-                                                 uint8_t hops, uint32_t slot) {
+                                                 uint16_t hops, uint32_t slot) {
     MOZ_ASSERT(slot < ENVCOORD_SLOT_LIMIT);
     return NameLocation(Kind::DebugEnvironmentCoordinate, bindKind, hops, slot);
   }
@@ -368,7 +368,7 @@ class NameLocation {
     return slot_;
   }
 
-  NameLocation addHops(uint8_t more) {
+  NameLocation addHops(uint16_t more) {
     MOZ_ASSERT(hops_ < ENVCOORD_HOPS_LIMIT - more);
     MOZ_ASSERT(kind_ == Kind::EnvironmentCoordinate);
     return NameLocation(kind_, bindingKind_, hops_ + more, slot_);
